Changes in lib/server/resources/operations
- changed patterns of Security Group names to prefix with 'safescale-sg_' (easier to find in list of Security Groups at provider side)
- reviewed and refined Security Group Rumes for Siubnet's internal Security Group (making it works as wanted in AWS)

+++ lib/server/resources/operations/converters/abstract.go
@@ -357,8 +357,8 @@ func SecurityGroupRuleFromAbstractToProtocol(in abstract.SecurityGroupRule) *pro
 		Direction:   protocol.SecurityGroupRuleDirection(in.Direction),
 		Protocol:    in.Protocol,
 		EtherType:   protocol.SecurityGroupRuleEtherType(in.EtherType),
-		PortFrom:    uint32(in.PortFrom),
-		PortTo:      uint32(in.PortTo),
+		PortFrom:    in.PortFrom,
+		PortTo:      in.PortTo,
 		IpRanges:    in.IPRanges,
 	}
 	return out
